frameworks/smolagents/requirements.txt
smolagents
python-dotenv
langchain-core
langchain-community
langchain-ollama
faiss-cpu
pypdf
mlx-lm
llama-cpp-python
openai
